fix file overwrite, texts fixed width 80

+++ deployqtmacx.sh
@@ -569,6 +569,14 @@ copy_dependency () {
 				local filename=$(basename "$srcpath")
 				filename="${filename%%.framework}"
 
+				if [[ -e "$dstpath" ]]; then
+					if [[ "$_FORCE_OVERWRITE" == true ]];
+						rm -Rf "$dstpath"
+					else
+						return 0
+					fi
+				fi
+
 				mkdir -p "$dstpath/Versions/A"
 
 				cp "$srcpath/Versions/A/$filename" "$dstpath/Versions/A/$filename"
@@ -578,10 +586,14 @@ copy_dependency () {
 				ln -s "Versions/A/$filename" "$filename"
 				ln -s "Versions/A" "Versions/Current"
 				cd "$pwd"
-			else
+			elif [[ "$_FORCE_OVERWRITE" == true ]];
+				cp -Rf "$srcpath" "$dstpath"
+			elif [[ ! -e "$dstpath" ]]
 				cp -R "$srcpath" "$dstpath"
 			fi
-		else
+		elif [[ "$_FORCE_OVERWRITE" == true ]]; then
+			cp -f "$srcpath" "$dstpath"
+		elif [[ ! -e "$$dstpath" ]]; then
 			cp "$srcpath" "$dstpath"
 		fi
 	elif [[ "$_VERBOSE" == true ]]; then
